オリジナルのAPIからの変更・非互換のためのプロパティを属性に追加。URIを返す戻り値の型をstringからUriに変更。

+++ NJaneScript/NJaneScriptBase/Wrapper/CosSim.cs
@@ -35,8 +35,8 @@ namespace NJaneScript.Wrapper {
 		/// <param name="Str">Subjectと比較する文字列</param>
 		/// <returns>SubjectとStrの類似度</returns>
 		[JaneScriptApi]
-		public float Evaluate(string Str) {
-			return (float)base.InvokeMethod("Evaluate", Str);
+		public double Evaluate(string Str) {
+			return (double)base.InvokeMethod("Evaluate", Str);
 		}
 	}
 }
